LCD working.

+++ src/ctrl.h
@@ -10,26 +10,26 @@
 
 typedef struct Ctrl Ctrl;
 struct Ctrl {
-        /* deviation */
-        int e;
-        /* sum of deviations */
-        int e_sum;
-        /* setpoint */
-        int sp;
-        /* period */
-        int period;
-        /* start time */
-        int t0;
-        /* end time */
-        int t1;
-        /* min pulse width */
-        int t1_min;
-        /* action max threshold */
-        int T_act;
-        /* P const. */
-        int P;
-        /* I const. */
-        int I;
+	/* deviation */
+	int e;
+	/* sum of deviations */
+	int e_sum;
+	/* setpoint */
+	int sp;
+	/* period */
+	long period;
+	/* start time */
+	long t0;
+	/* end time */
+	long t1;
+	/* min pulse width */
+	int t1_min;
+	/* action max threshold */
+	int T_act;
+	/* P const. */
+	int P;
+	/* I const. */
+	int I;
 };
 
 void ctrl_init(Ctrl *c);
